refactor: break up tui/mod.rs into operations + sessions + paths
Extracts standalone functions from the 4000-line src/tui/mod.rs into
cohesive sibling modules:

- src/tui/worktree_ops.rs — stale filter, delete, WorktreeSnapshot
- src/tui/sessions.rs — main/standalone session creation, collision detection
- src/paths.rs — sanitize_branch_slug, derive_local_worktree_path (promoted
  from tui-private to top-level domain module)

Also:
- Inlined derive_from_all_caches passthrough wrapper.
- Dedupped delete_worktree(&Worktree) into delete_task_row(&WorktreeRow);
  DeleteState.target now carries WorktreeRow end-to-end, eliminating the
  list::worktree_from_task_row conversion helper.
- Tightened visibility (pub(crate) → pub(super) where cross-module callers
  don't exist).
- Clarified canonicalize() intent in derive_local_worktree_path.
- Explicit named imports replaced glob use operations::*.

Refs #107
